bb-config: Add remote-configs field
Since we already support extending complete configs (not just single Os
List Item), provide a way to supply remote config URLs in the config
itself.

This is useful for times when the limitation of OS List (mainly the
single flasher for the whole list) can be a problem. This recently came
up with MicroPython

Signed-off-by: Ayush Singh <ayush@beagleboard.org>